requirements.txt
# MongoDB driver
pymongo>=4.0.0

# Environment variable management
python-dotenv>=1.0.0
